+++ packages/backend/src/queryBuilder.test.ts
@@ -1,17 +1,18 @@
-import { BinType, ForbiddenError, WeekDay } from '@lightdash/common';
+import {
+    BinType,
+    CustomDimensionType,
+    ForbiddenError,
+    isCustomBinDimension,
+} from '@lightdash/common';
 import {
     assertValidDimensionRequiredAttribute,
     buildQuery,
-    getCustomDimensionSql,
+    getCustomBinDimensionSql,
     replaceUserAttributes,
-    sortDayOfWeekName,
-    sortMonthName,
 } from './queryBuilder';
 import {
     bigqueryClientMock,
     COMPILED_DIMENSION,
-    COMPILED_MONTH_NAME_DIMENSION,
-    COMPILED_WEEK_NAME_DIMENSION,
     EXPLORE,
     EXPLORE_ALL_JOIN_TYPES_CHAIN,
     EXPLORE_BIGQUERY,
@@ -55,9 +56,7 @@ import {
     METRIC_QUERY_WITH_TABLE_CALCULATION_FILTER_SQL,
     METRIC_QUERY_WITH_TABLE_REFERENCE,
     METRIC_QUERY_WITH_TABLE_REFERENCE_SQL,
-    MONTH_NAME_SORT_SQL,
     warehouseClientMock,
-    WEEK_NAME_SORT_SQL,
 } from './queryBuilder.mock';
 
 describe('Query builder', () => {
@@ -516,10 +515,10 @@ describe('assertValidDimensionRequiredAttribute', () => {
 describe('with custom dimensions', () => {
     it('getCustomDimensionSql with empty custom dimension', () => {
         expect(
-            getCustomDimensionSql({
+            getCustomBinDimensionSql({
                 warehouseClient: bigqueryClientMock,
                 explore: EXPLORE,
-                compiledMetricQuery: METRIC_QUERY,
+                customDimensions: undefined,
                 userAttributes: {},
                 sorts: [],
             }),
@@ -528,11 +527,14 @@ describe('with custom dimensions', () => {
 
     it('getCustomDimensionSql with custom dimension', () => {
         expect(
-            getCustomDimensionSql({
+            getCustomBinDimensionSql({
                 warehouseClient: bigqueryClientMock,
 
                 explore: EXPLORE,
-                compiledMetricQuery: METRIC_QUERY_WITH_CUSTOM_DIMENSION,
+                customDimensions:
+                    METRIC_QUERY_WITH_CUSTOM_DIMENSION.customDimensions?.filter(
+                        isCustomBinDimension,
+                    ),
                 userAttributes: {},
                 sorts: [],
             }),
@@ -563,23 +565,21 @@ ELSE CONCAT(age_range_cte.min_id + age_range_cte.bin_width * 2, ' - ', age_range
 
     it('getCustomDimensionSql with only 1 bin', () => {
         expect(
-            getCustomDimensionSql({
+            getCustomBinDimensionSql({
                 warehouseClient: bigqueryClientMock,
 
                 explore: EXPLORE,
-                compiledMetricQuery: {
-                    ...METRIC_QUERY_WITH_CUSTOM_DIMENSION,
-                    customDimensions: [
-                        {
-                            id: 'age_range',
-                            name: 'Age range',
-                            dimensionId: 'table1_dim1',
-                            table: 'table1',
-                            binType: BinType.FIXED_NUMBER,
-                            binNumber: 1,
-                        },
-                    ],
-                },
+                customDimensions: [
+                    {
+                        id: 'age_range',
+                        name: 'Age range',
+                        type: CustomDimensionType.BIN,
+                        dimensionId: 'table1_dim1',
+                        table: 'table1',
+                        binType: BinType.FIXED_NUMBER,
+                        binNumber: 1,
+                    },
+                ],
                 userAttributes: {},
                 sorts: [],
             }),
@@ -647,6 +647,7 @@ LIMIT 10`);
                         {
                             id: 'age_range',
                             name: 'Age range',
+                            type: CustomDimensionType.BIN,
                             dimensionId: 'table1_dim1',
                             table: 'table1',
                             binType: BinType.FIXED_WIDTH,
@@ -733,11 +734,13 @@ LIMIT 10`);
 
     it('getCustomDimensionSql with sorted custom dimension ', () => {
         expect(
-            getCustomDimensionSql({
+            getCustomBinDimensionSql({
                 warehouseClient: bigqueryClientMock,
-
                 explore: EXPLORE,
-                compiledMetricQuery: METRIC_QUERY_WITH_CUSTOM_DIMENSION,
+                customDimensions:
+                    METRIC_QUERY_WITH_CUSTOM_DIMENSION.customDimensions?.filter(
+                        isCustomBinDimension,
+                    ),
                 userAttributes: {},
                 sorts: [{ fieldId: 'age_range', descending: true }],
             }),
@@ -829,6 +832,7 @@ LIMIT 10`);
                         {
                             id: 'age_range',
                             name: 'Age range',
+                            type: CustomDimensionType.BIN,
                             dimensionId: 'table1_dim1',
                             table: 'table1',
                             binType: BinType.FIXED_WIDTH,